+++ site/changelog.md
@@ -6,7 +6,7 @@
 
 - ProLayout
   - 🛠 `BaseMenu` 将原 `MenuUtil` 类改为纯函数构建 `NavMenuNode`，去掉为适配 antd `Menu` 遗留的 class 形态
-  - 🛠 `BaseMenu` 拆分 `baseMenuItemPath.tsx`：`getMenuItemPath`、外链分支与收起态 `MenuItemTooltip` 独立文件
+  - 🛠 `BaseMenu` 拆分 `baseMenuNavNodes.tsx`：`getIntlMenuName`、`getSubMenuOrItem`、`getMenuItemPath` 与收起态 `MenuItemTooltip` 独立文件
   - 🗑 移除 `layout="mix"` 类型；运行时仍接受字符串 `mix` 并映射为 `side`。原「顶栏一级 + 侧栏子级」请使用 `layout="side"` 与 `splitMenus`，顶栏菜单可用 `headerMenuRender` 自定义
   - 🗑 移除 `navTheme`（`ProSettings` / 路由 `MenuDataItem` 等不再透传）；侧栏 `Sider` 固定为 `theme="light"`，全局暗色请使用 antd `ConfigProvider` 的 `theme.algorithm`
   - 🗑 移除内置 `SettingDrawer`（不再从 `@ant-design/pro-components` 导出）
