+++ src/layout/components/SiderMenu/BaseMenu.tsx
@@ -20,14 +20,12 @@ import type {
   WithFalse,
 } from '../../typing';
 import { getOpenKeysFromMenuData } from '../../utils/utils';
-import type { PrivateSiderMenuProps } from './SiderMenu';
 import type { NavMenuNode } from './navMenuTypes';
+import type { PrivateSiderMenuProps } from './SiderMenu';
 import {
-  getIntlMenuName,
-  getMenuItemPath,
-  getMenuTitleSymbol,
+  getSubMenuOrItem,
   type MenuNavBuildContext,
-} from './baseMenuItemPath';
+} from './baseMenuNavNodes';
 import { ProLayoutNavMenu } from './ProLayoutNavMenu';
 import { useStyle } from './style/menu';
 import type {
@@ -153,133 +151,6 @@ const getIcon = (
   }
   return icon;
 };
-
-function getSubMenuOrItem(
-  ctx: MenuNavBuildContext,
-  item: MenuDataItem,
-  level: number,
-  noGroupLevel: number,
-): NavMenuNode | NavMenuNode[] {
-  const {
-    subMenuItemRender,
-    baseClassName,
-    collapsed,
-    menu,
-    iconPrefixes,
-    layout,
-    hashId,
-  } = ctx;
-  const isGroup = menu?.type === 'group' && layout !== 'top';
-  const name = getIntlMenuName(ctx, item);
-  const children = item?.children;
-
-  const menuType = isGroup && level === 0 ? ('group' as const) : undefined;
-
-  if (Array.isArray(children) && children.length > 0) {
-    const shouldHasIcon = level === 0 || (isGroup && level === 1);
-
-    const iconDom = getIcon(
-      item.icon,
-      iconPrefixes,
-      `${baseClassName}-icon ${hashId}`,
-    );
-    const defaultIcon =
-      collapsed && shouldHasIcon ? getMenuTitleSymbol(name) : null;
-
-    const defaultTitle = (
-      <div
-        className={clsx(`${baseClassName}-item-title`, hashId, {
-          [`${baseClassName}-item-title-collapsed`]: collapsed,
-          [`${baseClassName}-item-title-collapsed-level-${noGroupLevel}`]:
-            collapsed,
-          [`${baseClassName}-group-item-title`]: menuType === 'group',
-          [`${baseClassName}-item-collapsed-show-title`]:
-            menu?.collapsedShowTitle && collapsed,
-        })}
-      >
-        {menuType === 'group' && collapsed ? null : shouldHasIcon && iconDom ? (
-          <span className={clsx(`${baseClassName}-item-icon`, hashId)}>
-            {iconDom}
-          </span>
-        ) : (
-          defaultIcon
-        )}
-        <span
-          className={clsx(`${baseClassName}-item-text`, hashId, {
-            [`${baseClassName}-item-text-has-icon`]:
-              menuType !== 'group' &&
-              shouldHasIcon &&
-              (iconDom || defaultIcon),
-          })}
-        >
-          {name}
-        </span>
-      </div>
-    );
-
-    const title = subMenuItemRender
-      ? subMenuItemRender({ ...item, isUrl: false }, defaultTitle, ctx)
-      : defaultTitle;
-
-    const nextNoGroupLevel =
-      isGroup && level === 0 && ctx.collapsed ? level : level + 1;
-    const childrenList = children
-      .map((child) => getSubMenuOrItem(ctx, child, level + 1, nextNoGroupLevel))
-      .filter((item) => item)
-      .flat(1) as NavMenuNode[];
-
-    const submenuOrGroup: NavMenuNode =
-      menuType === 'group'
-        ? {
-            kind: 'group',
-            key: String(item.key! || item.path!),
-            label: title,
-            children: childrenList,
-            className: clsx(`${baseClassName}-group`),
-          }
-        : {
-            kind: 'submenu',
-            key: String(item.key! || item.path!),
-            label: title,
-            onTitleClick: (e) => item.onTitleClick?.(e),
-            children: childrenList,
-            hasIcon: !!(shouldHasIcon && iconDom),
-            className: clsx({
-              [`${baseClassName}-submenu`]: true,
-              [`${baseClassName}-submenu-has-icon`]: shouldHasIcon && iconDom,
-            }),
-          };
-
-    return [
-      submenuOrGroup,
-      isGroup && level === 0
-        ? ({
-            kind: 'divider' as const,
-            key: `${item.key! || item.path!}-group-divider`,
-            className: `${baseClassName}-divider`,
-            style: {
-              padding: 0,
-              borderBlockEnd: 0,
-              margin: ctx.collapsed ? '4px' : '6px 16px',
-              marginBlockStart: ctx.collapsed ? 4 : 8,
-              borderColor: 'var(--pro-layout-nav-color-divider)',
-            },
-          } as NavMenuNode)
-        : undefined,
-    ].filter(Boolean) as NavMenuNode[];
-  }
-
-  const onTitle = (item as MenuDataItem & { onTitleClick?: () => void })
-    .onTitleClick;
-  return {
-    kind: 'item' as const,
-    className: `${baseClassName}-menu-item`,
-    disabled: item.disabled,
-    key: String(item.key! || item.path!),
-    onClick: onTitle ? () => onTitle() : undefined,
-    label: getMenuItemPath(ctx, item, level, noGroupLevel, getIcon),
-  };
-}
 
 const BaseMenu: React.FC<BaseMenuProps & PrivateSiderMenuProps> = (props) => {
   const { getPrefixCls } = useContext(ConfigProvider.ConfigContext);
@@ -503,7 +374,7 @@ const BaseMenu: React.FC<BaseMenuProps & PrivateSiderMenuProps> = (props) => {
       openKeys={navOpenKeys}
       defaultOpenKeys={defaultOpenKeysRef.current.map((k) => String(k))}
       nodes={(finallyData ?? [])
-        .map((item) => getSubMenuOrItem(menuNavCtx, item, 0, 0))
+        .map((item) => getSubMenuOrItem(menuNavCtx, item, 0, 0, getIcon))
         .filter(Boolean)
         .flat(1) as NavMenuNode[]}
       onOpenChange={(_openKeys) => {
